fix(swap): send KyberSwap slippage tolerance as 100 bps
KyberSwap denominates slippageTolerance in basis points (10000 = 100%).
Android sent 2.5 (~0.025%), two orders of magnitude tighter than the 100
(1%) that iOS and the extension send, so Kyber route builds and on-chain
swaps reverted for most pairs. Send 100 on both the routes query and the
route/build body, and type the request field as Int so the build POST
serialises a clean 100 instead of 100.0.

+++ data/src/main/kotlin/com/vultisig/wallet/data/api/models/KyberJson.kt
@@ -73,7 +73,7 @@ data class KyberSwapBuildRequest(
     val routeSummary: KyberSwapRouteResponse.RouteSummary,
     val sender: String,
     val recipient: String,
-    val slippageTolerance: Double,
+    val slippageTolerance: Int,
     val deadline: Int,
     val enableGasEstimation: Boolean,
     val source: String,
